% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_email_for_potential_reviewer.R
\name{draft_email_for_potential_reviewer}
\alias{draft_email_for_potential_reviewer}
\title{Draft email for a potential reviewer}
\usage{
draft_email_for_potential_reviewer(
  reviewer_name = "Jane Smith",
  package = "coolr",
  author_name = "John Doe",
  package_does = "does cool things in R",
  package_repo_url = "https://github.com/jhondoe/coolr",
  package_review_issue_url = "https://github.com/ropensci/software-review/issues/000",
  why_reviewer_fits = "of your experience developing cool things"
)
}
\arguments{
\item{reviewer_name, package, author_name, package_does, package_repo_url, package_review_issue_url, why_reviewer_fits}{Character.}
}
\description{
Draft email for a potential reviewer
}
\examples{
draft_email_for_potential_reviewer()
}
