% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_template.R
\name{editor_template}
\alias{editor_template}
\title{Print the editor's template}
\usage{
editor_template()
}
\value{
Text.
}
\description{
\url{https://devdevguide.netlify.app/editortemplate}
}
\examples{
editor_template()
}
