% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic_check.R
\name{eic_check}
\alias{eic_check}
\alias{app_eic_check}
\title{Editor (pre) checks}
\usage{
eic_check(handle = "johndoe")

app_eic_check()
}
\arguments{
\item{handle}{Character. The handle of the package author.}
}
\value{
Text
}
\description{
As the EiC create the first comment following a (pre) submission, including
a template for preliminary editor checks.
}
\examples{
eic_check()

if (interactive()) app_eic_check()
}
