% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_reviewer.R
\name{email_reviewer}
\alias{email_reviewer}
\title{Draft email for a potential reviewer}
\usage{
email_reviewer(
  reviewer_name = "Jane Smith",
  why_reviewer_fits = "of your experience developing cool things",
  package = "coolr",
  author_name = "John Doe",
  package_does = "does cool things in R",
  package_repo_url = "https://github.com/jhondoe/coolr",
  package_review_issue_url = "https://github.com/ropensci/software-review/issues/000"
)
}
\arguments{
\item{reviewer_name}{Reviewer's name.}

\item{why_reviewer_fits}{Reason why the reviewer is a good fit.}

\item{package}{Name of the package.}

\item{author_name}{Author's name.}

\item{package_does}{What the package does.}

\item{package_repo_url}{URL of the package repository.}

\item{package_review_issue_url}{URL of the package review issue.}
}
\value{
Character. The generated email content.
}
\description{
Draft email for a potential reviewer
}
\examples{
email_reviewer()
}
