% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Get an issue from ropensci/software-review}
\usage{
get_comments(number)
}
\arguments{
\item{number}{The issue number}
}
\value{
A list of the issue
}
\description{
Get an issue from ropensci/software-review
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_comments(688)
\dontshow{\}) # examplesIf}
}
