% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invite_full_submission.R
\name{invite_full_submission}
\alias{invite_full_submission}
\alias{app_invite_full_submission}
\title{Invite full submission}
\usage{
invite_full_submission(
  handle = "johndoe",
  guide_url = c("https://devdevguide.netlify.app/softwarereview_author",
    "https://stats-devguide.ropensci.org/pkgdev.html")
)

app_invite_full_submission()
}
\arguments{
\item{handle}{Character. The handle of the package author.}

\item{guide_url}{Character. The URL of the guide for authors.}
}
\value{
Text
}
\description{
As an EiC invite a full submission.
}
\examples{
invite_full_submission()

if (interactive()) app_invite_full_submission()
}
