% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropensciutils-package.R
\docType{package}
\name{ropensciutils-package}
\alias{ropensciutils}
\alias{ropensciutils-package}
\title{ropensciutils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
