% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack_editors.R
\name{slack_editors}
\alias{slack_editors}
\title{Generate email to editors for a new package submission}
\usage{
slack_editors(
  package_review_issue_url = "https://github.com/ropensci/software-review/issues/000",
  package_repo_url = "https://github.com/jhondoe/coolr",
  type = "standard",
  category = "category 'workflow automation'",
  package = "coolr",
  package_does = "does cool things in R",
  package_check_shows = "this package is in top shape",
  in_scope_because = "it streamlines necessary processes",
  my_checks_show = "the package is awesome",
  my_checks_url =
    "https://github.com/ropensci/software-review/issues/000#issuecomment-0000000000",
  category_description =
    "Tools that automate and link together workflows, such as build systems and tools to manage continuous integration. Does not include general tools for literate programming. (e.g., R markdown extensions not under the previous topics). (Example: drake)"
)
}
\arguments{
\item{package_review_issue_url}{URL of the package review issue.}

\item{package_repo_url}{URL of the package repository.}

\item{type}{Type of submission (e.g., "standard").}

\item{category}{Category of the package (e.g., "workflow automation").}

\item{package}{Name of the package.}

\item{package_does}{What the package does.}

\item{package_check_shows}{Summary of the package check results.}

\item{in_scope_because}{Reason why the package is in scope.}

\item{my_checks_show}{Summary of the preliminary checks performed.}

\item{my_checks_url}{URL to the detailed preliminary checks.}

\item{category_description}{Description of the category.}
}
\value{
Character. The generated email content.
}
\description{
Generate email to editors for a new package submission
}
\examples{
slack_editors()
}
