% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/software_review_issue.R
\name{software_review_issue}
\alias{software_review_issue}
\alias{software_review_chat}
\title{Chat about a specific software review issue}
\usage{
software_review_issue(issue)

software_review_chat(chat, prompt)
}
\arguments{
\item{issue}{An integer representing the issue number. This identifies the specific issue for which comments should be retrieved.}

\item{chat}{An ellmer chat object}

\item{prompt}{Character. A prompt to chat about a specific software review issue.}
}
\value{
A character string containing a message indicating where to look
for answers, along with a JSON representation of the comments associated
with the specified issue.
}
\description{
Creates the system prompt required to focus a chat to a specific
issue in ropensci/software-review.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
chat <- software_review_issue(688) |> chat_gemini()
chat |> software_review_chat("what did maurolepore say?")
\dontshow{\}) # examplesIf}
}
